backend/requirements.txt
fastapi~=0.104
uvicorn~=0.24
google-generativeai~=0.3
pydantic~=2.5
python-dotenv~=1.0
